requirements.txt
# pip install -r requirements.txt

Cython
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
Pillow
PyYAML>=5.3
scipy>=1.4.1
torch>=1.9.1
torchvision>=0.10.1 
tqdm>=4.41.0
seaborn
pandas
timm
